chore: refactor and package modules
