+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include README.md
+include LICENSE
+recursive-include src/pinet *.json *.yaml
