+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-testpaths = test
+testpaths = src/test
 pythonpath = .
