+++ src/benchmarks/configs/benchmark_cvxpy.yaml
