+++ src/benchmarks/configs/benchmark_jaxopt_small.yaml
