+++ src/benchmarks/configs/benchmark_softpen_small.yaml
