+++ src/benchmarks/configs/toy_MPC.yaml
