--- src/hcnn/constraints/__init__.py
@@ -1,1 +1,0 @@
-"""Constraints module for the HCNN package."""
