--- src/hcnn/solver/__init__.py
@@ -1,1 +1,0 @@
-"""Solver module for the HCNN package."""
