+++ src/pinet/constraints/affine_inequality.py
@@ -2,8 +2,9 @@
 
 from jax import numpy as jnp
 
-from hcnn.constraints.base import Constraint
-from hcnn.utils import Inputs
+from pinet.dataclasses import Inputs
+
+from .base import Constraint
 
 
 class AffineInequalityConstraint(Constraint):
