+++ src/pinet/solver/__init__.py
@@ -1,0 +1,8 @@
+"""Solver module for the HCNN package."""
+
+from .admm import build_iteration_step
+
+__all__ = [
+    "build_iteration_step",
+    "Inputs",
+]
