+++ src/test/test_vjp.py
@@ -9,11 +9,13 @@ import jax.test_util
 import pytest
 from jax import numpy as jnp
 
-from hcnn.constraints.affine_equality import EqualityConstraint
-from hcnn.constraints.affine_inequality import AffineInequalityConstraint
-from hcnn.constraints.box import BoxConstraint
-from hcnn.project import Project
-from hcnn.utils import Inputs
+from pinet import (
+    AffineInequalityConstraint,
+    BoxConstraint,
+    EqualityConstraint,
+    Inputs,
+    Project,
+)
 
 jax.config.update("jax_enable_x64", True)
